+++ src/Form/PlanningType.php
@@ -13,7 +13,6 @@ class PlanningType extends AbstractType
     {
         $builder
             ->add('dataProvider')
-            ->add('viewType')
         ;
     }